requirements.txt
opencv-python
PyQt5
Pillow
scikit-learn
numpy
matplotlib
scipy
guppy3